+++ app/routes/people.rb
@@ -15,5 +15,12 @@ module SalesLoft
                 @users = SalesLoftController.instance.people_available
             end
         end
+
+        resource :frequency do
+            desc 'Show the frequency map of all characters present in people email addresses'
+            get erb: 'people/frequency' do
+                @frequency_map = []
+            end
+        end
     end
 end