+++ ReleaseNotes.md
@@ -3,3 +3,4 @@
 # Level 1
  - [1] Application scaffolding.
  - [2] Create endpoint to show available People list.
+ - [3] Add a button that shows a frequency map of people's email characters.